# JIT-compile the ODE time step functions