--- a/test/dynamics/test_driver.py
+++ b/test/dynamics/test_driver.py
@@ -26,7 +26,7 @@ import netket.experimental as nkx
 SEED = 214748364
 
 
-def _setup_system(L, *, dtype=np.float64):
+def _setup_system(L, *, dtype=np.complex128):
     g = nk.graph.Chain(length=L)
     hi = nk.hilbert.Spin(s=0.5, N=g.n_nodes)
 
@@ -84,7 +84,7 @@ def test_one_fixed_step(integrator, propagation_type):
     assert te.t == 0.01
 
 
-def l4_norm(_, x):
+def l4_norm(x):
     """
     Custom L4 error norm.
     """
@@ -229,4 +229,4 @@ def test_repr_and_info():
     assert "TDVP" in info
     assert "generator" in info
     assert "integrator" in info
-    assert "rk23" in info
+    assert "RK23" in info