@femtobit 2021-12-09
The remaining test failure is strange:
```
 worker 'gw0' crashed while running 'test/variational/test_variational.py::test_expect[model:(C->C)-operator:(Hermitian Complex)]'
```
I didn't see that locally. Could this be another consequence of the JAX update?